+++ src/platform_impl/linux/wayland/window.rs
@@ -2,6 +2,7 @@ use std::{
     collections::VecDeque,
     sync::{Arc, Mutex, Weak},
 };
+use raw_window_handle::unix::WaylandHandle;
 
 use crate::{
     dpi::{LogicalPosition, LogicalSize},
@@ -332,6 +333,14 @@ impl Window {
 
     pub fn primary_monitor(&self) -> MonitorHandle {
         primary_monitor(&self.outputs)
+    }
+
+    pub fn raw_window_handle(&self) -> WaylandHandle {
+        WaylandHandle {
+            surface: self.surface.c_ptr() as *mut _,
+            display: self.display.c_ptr() as *mut _,
+            ..WaylandHandle::empty()
+        }
     }
 }