+++ src/platform_impl/linux/x11/window.rs
@@ -9,6 +9,7 @@ use std::{
     ptr, slice,
     sync::Arc,
 };
+use raw_window_handle::unix::X11Handle;
 
 use libc;
 use parking_lot::Mutex;
@@ -1374,5 +1375,14 @@ impl UnownedWindow {
             .lock()
             .unwrap()
             .insert(WindowId(self.xwindow));
+    }
+
+    #[inline]
+    pub fn raw_window_handle(&self) -> X11Handle {
+        X11Handle {
+            window: self.xwindow,
+            display: self.xconn.display as _,
+            ..X11Handle::empty()
+        }
     }
 }