+++ src/platform_impl/macos/window.rs
@@ -7,6 +7,7 @@ use std::{
         Arc, Mutex, Weak,
     },
 };
+use raw_window_handle::{RawWindowHandle, macos::MacOSHandle};
 
 use crate::{
     dpi::{LogicalPosition, LogicalSize},
@@ -903,6 +904,16 @@ impl UnownedWindow {
     #[inline]
     pub fn primary_monitor(&self) -> MonitorHandle {
         monitor::primary_monitor()
+    }
+
+    #[inline]
+    pub fn raw_window_handle(&self) -> RawWindowHandle {
+        let handle = MacOSHandle {
+            ns_window: self.ns_window as *mut _,
+            ns_view: self.ns_view as *mut _,
+            ..MacOSHandle::empty()
+        };
+        RawWindowHandle::MacOS(handle)
     }
 }